// src/app/models/titular_renovacion.model.ts
export class TitularRenovacionModel {
    comiteRenovacionId?: number;
    titularId?: number;
    tipoRenovacion?: string;
    flagRenovacion?: number;
    resolucion?: string;
    resolucionFecha?: string;
    resolucionVigencia?: number;
    vigenciaDesde?: string;
    vigenciaHasta?: string;
    estado?: number;
    usuarioRegistro?: number;
    accion?: string;
}